refactor : mongoDB 데이터 저장 관련 로직 MongoDBManager 클래스로 이전
