+++ data-ingestor/src/main/java/com/hanyang/dataingestor/infrastructure/DeadLetterQueueConsumer.java
@@ -39,7 +39,6 @@ public class DeadLetterQueueConsumer {
             log.warn("실패 시간: {}", headers.get("x-failure-time"));
             log.warn("에러 메시지: {}", headers.get("x-error-message"));
 
-            mongoManager.dropIfExists(messageDto.getDatasetId());
             dataParsingService.createDataTable(messageDto.getDatasetId());
             s3StorageManager.deleteDatasetFiles(messageDto.getDatasetId());
             channel.basicAck(tag, false);
